Move pandas styler logic to dedicated module
